pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyllo"
version = "0.1.0"
description = "Clay-science retrieval augmented generation toolkit."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "pymupdf>=1.23.26",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sentence-transformers>=2.2.2,<3.0",
    "transformers>=4.30,<4.38",
    "faiss-cpu>=1.7.4",
    "numpy>=1.23.0",
    "pandas>=1.5.0",
    "openai>=1.6.0",
    "litellm>=1.35.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = ["black", "ruff", "mypy", "pytest"]

[project.scripts]
pyllo = "pyllo.cli:main"

[tool.typer]
name = "pyllo"
